Automation_Bot/requirements.txt
flask
stripe
selenium
webdriver-manager
python-dotenv
pandas
openpyxl